last commit

+++ app/Http/Controllers/CheckListItemController.php
@@ -7,66 +7,28 @@ use App\CheckListItem;
 use Illuminate\Http\Request;
 
 class CheckListItemController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     *
-     * @param  \App\CheckList  $checkList
-     * @return \Illuminate\Http\Response
-     */
-    public function index(CheckList $checkList)
-    {
-        //
-    }
-
+{    
     /**
      * Show the form for creating a new resource.
      *
      * @param  \App\CheckList  $checkList
      * @return \Illuminate\Http\Response
      */
-    public function create(CheckList $checkList)
+    /* public function create(CheckList $checkList)
     {
+        $this->authorize('create', $checkList);
         $item = new CheckListItem();
         $checkList->items()->save($item);
         $checkList->save();
         return redirect()->route('check_lists.show', $checkList);
-    }
+    } */
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\CheckList  $checkList
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request, CheckList $checkList)
+    public function store(CheckList $checkList)
     {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\CheckList  $checkList
-     * @param  \App\CheckListItem  $checkListItem
-     * @return \Illuminate\Http\Response
-     */
-    public function show(CheckList $checkList, CheckListItem $checkListItem)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\CheckList  $checkList
-     * @param  \App\CheckListItem  $checkListItem
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(CheckList $checkList, CheckListItem $checkListItem)
-    {
-        //
+        $this->authorize('update', $checkList);
+        $item = new CheckListItem();
+        $checkList->items()->save($item);
+        return redirect()->route('check_lists.show', $checkList);
     }
 
     /**
@@ -79,7 +41,7 @@ class CheckListItemController extends Controller
      */
     public function update(Request $request, CheckList $checkList, CheckListItem $checkListItem)
     {
-        #dd($request->all());
+        $this->authorize('update', $checkListItem);
         $checkListItem->text = $request->text;
         $checkListItem->status = $request->status ?? false;
         $checkListItem->save();
@@ -96,6 +58,7 @@ class CheckListItemController extends Controller
      */
     public function destroy(CheckList $checkList, CheckListItem $checkListItem)
     {
+        $this->authorize('delete', $checkListItem);
         $checkListItem->delete();
         flash(__('Deleted'))->success();
         return redirect()->route('check_lists.show', $checkList);
